Move BotWrapper into bot_wrapper.rs
